Add form error handling

+++ src/main/java/com/tomdud/businesstripapp/servlet/admin/ReceiptTypeServlet.java
@@ -1,5 +1,6 @@
 package com.tomdud.businesstripapp.servlet.admin;
 
+import com.tomdud.businesstripapp.exception.FormValueNotCorrectException;
 import com.tomdud.businesstripapp.model.ReceiptType;
 import com.tomdud.businesstripapp.service.ReceiptTypeService;
 import com.tomdud.businesstripapp.servlet.calculator.CalculateReimbursementServlet;
@@ -24,26 +25,37 @@ public class ReceiptTypeServlet extends HttpServlet {
 
         logger.log(Level.INFO, "ReceiptTypeServlet::doGet - action - {0}", action);
 
-        switch (action) {
-            case "/receipt-type-delete":
-                String receiptNameToDelete = request.getParameter("name");
-                logger.log(Level.INFO, "ReceiptTypeServlet::doGet - deleting receipt with name {0}", receiptNameToDelete);
-                receiptTypeService.deleteReceiptTypeByName(receiptNameToDelete);
-                response.sendRedirect(request.getContextPath() + "/admin-panel");
-                break;
-            case "/receipt-type-edit":
-                String receiptNameToEdit = request.getParameter("name");
-                logger.log(Level.INFO, "ReceiptTypeServlet::doGet - edit receipt with name {0}", receiptNameToEdit);
-                ReceiptType receiptType = receiptTypeService.getReceiptTypeByName(receiptNameToEdit);
+        try {
+            switch (action) {
+                case "/receipt-type-delete":
+                    if (request.getParameter("name").length() == 0)
+                        throw new FormValueNotCorrectException("Provide correct receipt name");
 
-                request.setAttribute("receiptType", receiptType);
-                request.getRequestDispatcher("/receipt-type-edit.jsp").forward(request, response);
-                break;
-            case "/receipt-type-insert":
-                logger.log(Level.INFO, "ReceiptTypeServlet::doGet - inserting new receipt");
-                request.setAttribute("receiptType", null);
-                request.getRequestDispatcher("/receipt-type-edit.jsp").forward(request, response);
-                break;
+                    String receiptNameToDelete = request.getParameter("name");
+                    logger.log(Level.INFO, "ReceiptTypeServlet::doGet - deleting receipt with name {0}", receiptNameToDelete);
+                    receiptTypeService.deleteReceiptTypeByName(receiptNameToDelete);
+                    response.sendRedirect(request.getContextPath() + "/admin-panel");
+                    break;
+                case "/receipt-type-edit":
+                    if (request.getParameter("name").length() == 0)
+                        throw new FormValueNotCorrectException("Provide correct receipt name");
+
+                    String receiptNameToEdit = request.getParameter("name");
+                    logger.log(Level.INFO, "ReceiptTypeServlet::doGet - edit receipt with name {0}", receiptNameToEdit);
+                    ReceiptType receiptType = receiptTypeService.getReceiptTypeByName(receiptNameToEdit);
+
+                    request.setAttribute("receiptType", receiptType);
+                    request.getRequestDispatcher("/receipt-type-edit.jsp").forward(request, response);
+                    break;
+                case "/receipt-type-insert":
+                    logger.log(Level.INFO, "ReceiptTypeServlet::doGet - inserting new receipt");
+                    request.setAttribute("receiptType", null);
+                    request.getRequestDispatcher("/receipt-type-edit.jsp").forward(request, response);
+                    break;
+            }
+        } catch (Exception exception) {
+            response.sendRedirect(request.getContextPath() + "/admin-panel?error=true&errorMessage=" + exception.getMessage());
+            return;
         }
 
     }
@@ -54,25 +66,43 @@ public class ReceiptTypeServlet extends HttpServlet {
 
         logger.log(Level.INFO, "ReceiptTypeServlet::doPost - action - {0}", action);
 
-        switch (action) {
-            case "/receipt-type-update":
-                ReceiptType receiptType = receiptTypeService.getReceiptTypeByName(request.getParameter("orgName"));
-                receiptType.setName(request.getParameter("name"));
-                receiptType.setEnableLimit(request.getParameter("enableLimit") != null);
-                receiptType.setLimit(Double.parseDouble(request.getParameter("limit")));
+        try {
+            switch (action) {
+                case "/receipt-type-update":
+                    ReceiptType receiptType = receiptTypeService.getReceiptTypeByName(request.getParameter("orgName"));
 
-                response.sendRedirect(request.getContextPath() + "/admin-panel");
-                break;
-            case "/receipt-type-insert":
-                ReceiptType receiptTypeToSave = new ReceiptType(
-                        request.getParameter("name"),
-                        request.getParameter("enableLimit") != null,
-                        Double.parseDouble(request.getParameter("limit"))
-                );
-                receiptTypeService.saveReceiptType(receiptTypeToSave);
+                    if (request.getParameter("name").length() == 0)
+                        throw new FormValueNotCorrectException("Provide correct receipt name");
 
-                response.sendRedirect(request.getContextPath() + "/admin-panel");
-                break;
+                    if (request.getParameter("limit").length() == 0)
+                        throw new FormValueNotCorrectException("Provide correct receipt limit value");
+
+                    receiptType.setName(request.getParameter("name"));
+                    receiptType.setEnableLimit(request.getParameter("enableLimit") != null);
+                    receiptType.setLimit(Double.parseDouble(request.getParameter("limit")));
+
+                    response.sendRedirect(request.getContextPath() + "/admin-panel");
+                    break;
+                case "/receipt-type-insert":
+                    if (request.getParameter("name").length() == 0)
+                        throw new FormValueNotCorrectException("Provide correct receipt name");
+
+                    if (request.getParameter("limit").length() == 0)
+                        throw new FormValueNotCorrectException("Provide correct receipt limit value");
+
+                    ReceiptType receiptTypeToSave = new ReceiptType(
+                            request.getParameter("name"),
+                            request.getParameter("enableLimit") != null,
+                            Double.parseDouble(request.getParameter("limit"))
+                    );
+                    receiptTypeService.saveReceiptType(receiptTypeToSave);
+
+                    response.sendRedirect(request.getContextPath() + "/admin-panel");
+                    break;
+            }
+        } catch (Exception exception) {
+            response.sendRedirect(request.getContextPath() + "/admin-panel?error=true&errorMessage=" + exception.getMessage());
+            return;
         }
 
     }
